Don't explicitly initialize gcState
Make the AllocatedCell GC state have a value of zero and depend on the
fact that mmap() will give us zeroed memory. This speeds up large
allocations as the allocator doesn't need to pass through the memory
once just to have the same memory touched later when the type ID etc.
are set by the cell consumer.