Greatly simplify CellRefRangeList
Instead of allocating CellRefRange nodes out of a fixed size
DynamicMemoryBlock just embed it inside AbstractRangeRef. This means we
allocate the CellRefRange on the stack for free. This simplifies and
speeds the CellRefRangeList greatly

This is a first step to moving to codegen to its own thread safe shadow
stack implementation shared with the AbstractRef infrastructure.