runtime: Implicit data sharing using byte arrays
Note: This does not have matching codegen support. Codegen is completely
broken by this change.

This makes strings, symbols and bytevectors use implicitly shared byte
arrays when copying or transmuting entire values. This probably only has
a marginal memory usage benefit for most programs but lays the
groundwork for implicit sharing between worlds which makes inter-world
communication potentially a lot cheaper.

Allowing bytevectors to implicitly share with symbols and strings means
that symbols and strings are no longer implicitly NULL terminated. This
simplified the code somewhat and allows an extra character to be stored
inline. However, it means an additional StringCell method will need to
be introduced to return NULL terminated strings (probably as a
SharedByteArray)