runtime: Change memcpy => memmove in StringCell
This was originally memmove until the "sameString" slow case was added.
The idea was that if the source and destination are different strings
it's safe to use memcpy. However "sameString" refers to the destination
string and the *pattern* - it's still possible for the source and
destination string to be the same with sameString == false.