codegen: Fix constant symbols of exactly 12 bytes
Since the shared byte array changes the runtime thinks 12 bytes is the
symbol inline threshold while codegen was still using 11 bytes. This
means that the runtime would use the pointer address of 12 byte symbols
as the symbol's UTF-8 data.

Make codegen use the correct 12 byte limit and add a unit test for 12
byte symbols and strings.