stdlib: Add (unless), unit test for (when)
This uses the macro in R7RS but tweaks it and the (when) macro to always
return #!unit. R7RS says their result is unspecified but the included
macro will return the value of the expression if it's evaluated. Always
returning #!unit is consistent with how Llambda handles unspecified
values elsewhere.