runtime: Fast path for string and symbol ==
Check if the constUtf8Data() pointer is the same for both cells before
performing memcmp(). This means the cells are either the same cell or
implicitly sharing data.

This isn't likely to be true very often but the size check would reject
most inequalities early and we already need to calculate the pointer to
to the UTF-8 data. This means the test is effectively free assuming
we're cache miss limited here.