stdlib: Implement letrec*
R7RS has a macro definition for (letrec*) that assigns locations a
special "uninitialized" value and then implements multiple body
(define)s in terms of (letrec*)

We take the opposite approach of handling multiple body (define)s in
the compiler frontend and then uses a macro to implement (letrec*) in
terms of body defines. The result should be equivalent.