stdlib: Partially implement (letrec)
Implement (letrec) as a simple alias for (letrec*)

This isn't quite right.  It's legal to refer to values earlier in the
initializer list in (letrec*) but not in (letrec) This means all valid
(letrec) initializers are valid (letrec*) initializers but not the
converse.  We should be more strict with invalid (letrec) uses in the
future