Add list metrics for constant proper lists
This encodes the length and type of the list inside the pair cells for
constant lists and changes the runtime to check those fields when
creating a ProperList. This allows skipping a list scan for certain
operations and makes finding the length of a constant list O(1). The
downside is a slight slowdown for non-proper lists due to the extra
tests.