codegen: Properly phi non-results across branches
A condition's result was properly being phi'ed but any live temps
weren't. This means that if GC occurred in one of the branches the stale
value from before the branch would be used instead.

This resulted in unit test crashes with optimization disabled. However,
it's easily triggerable with -O 2 as the new functional test shows.