Run all functional tests at -O 0 and -O 2
The fact the functional tests always ran with optimisation enabled means
the unoptimised case was poorly tested and slowly bitrotting. When
additional optimisations are introduced this will only become more
problematic as the delta between the optimised and unoptimised code
increases and some trivial tests are optimised away completely.

This builds on the work of the previous 4 commits which fix tests which
were failing with optimisation disabled.