compiler: Basic support for inlining with rest arg
This only works if some simple operations (pair?, null? or length) are
applied to the rest arg or the reduced lambda body doesn't reference the
rest arg at all. This is the initial step to inlining (case-lambda).

This is implemented by introducing yet another abstract tree called a
PartialValue. PartialValues are similar to an expression tree except the
structure of an expression (pairs and vectors) can be known without the
contents of the structure being resolved to a literal. This allows
operations like (length) on a list where not all the members are
literals. This will also allow for re-applying a rest arg using (apply)