stdlib: Implement optional fill for (make-*)
Make the fill argument for (make-list), (make-vector) and
(make-bytevector) optional by using (case-lambda) in the stdlib. This
required moving the (case-lambda) definition much earlier in stdlib.scm
so it could be used by the rest of the base library.

It's been manually confirmed that direct invocations of the new
procedures will be properly reduced to direct native function calls at
-O 2. The code generated for the generalized (case-lambda) is suboptimal
but hopefully it won't be invoked too frequently.

(make-list) and (make-vector) default fill with unit values while
(make-bytevector) fill with zeros. R7RS allows us to fill with anything
if the fill is unspecified but those values seem reasonably sane.