compiler: Optimise GC rooting
Previously every GC-managed TempValue that passed through a GC barrier
was assigned a unique GC slot. This has a number of problems:

- TempValues created by casts will have identical pointer values and
  only one of them need to be saved

- Many TempValues have non-overlapping lifetimes and can therefore share
  the same slot space

Instead of tracking TempValues the GC code is concerned with "pointer
identities".  Normally every new TempValue is assigned a unique pointer
identity but casts will cause the new TempValue to point to the pointer
identity of its original. This solves the problem of identical pointers
being saved to the shadow stack

The overlapping lifetime issue is a bit trickier to solve. Firstly, we
now flush all of our unsaved roots to the shadow stack before entering a
conditional branch. This prevents branches from assigning new slots to
the same identity in their respective branches. Then, we keep track of
the next unallocated slot and the currently rooted slots. Any allocated
slot that isn't rooted is free to be reassigned to a newly rooting slot.

We have logic to prefer transferring slots that have just derooted to a
newly rooting slot. This means we can avoid nulling out the slot and can
just overwrite it directly.