compiler: Better tracking of which values to GC
When GC support was initially added to codegen TempValue had an
"isGcManaged" flag. Every user of TempValue needed to manually annotate
each new TempValue as either managed or unmanaged. This was nearly
manageable but we want to start considering all preconstructed values as
GC unmanaged which pushed the complexity over the edge

- Add TempValue constructors to each class of temporary value. This
  doesn't immediately contribute to a more strongly typed Step but it
  could become useful later.

- Convert nearly all users to the new TempValue constructors

- Have typegen give preconstructed cells the supertype of
  PreconstructedCellType

- Make ValueType consider preconstructed cells non-GC managed. This is
  the only (intentional) functional change which allows us to avoid
  rooting cell types we know can't have runtime allocated instances.