compiler: Detect side effectless apply
This adds reducer.reportproc.ProcHasSideEffects and
codegen.RuntimeFunctions.hasSideEffects. These conservatively determine
if the given report procedure or native function have side effects,
respectively. This is currenly done entirely based on the their name and
arity as the type information isn't fully known to the reducer. This
means any procedure with typed arguments must be considered to have side
effects as they may throw exceptions.

The motivation and end result of this is the port and parameter for
stdin, stdout and stderr are only created in programs that reference
them. This reduces GC overhead by avoiding up to 6 cell allocations
during early startup