compiler: Add _lliby_write_stdout for unit tests
Since the unit tests switched to using vanilla (write) they've been
greatly slowed down. They needed to generate the code for (write) and
its internal lambdas and well as generate at least two GC barriers for
the write itself. We don't really need the full power of (write) so
introduce a _lliby_write_stdout that requires no GC barriers. This is
used to output the last value of a test's expression

At the same time move _lliby_std{out,in,error}_port in to core/ as
they're not part of the Scheme standard library