compiler: Pop shadow stack before signalling error
If GC happens at the right time during exception handling it may try to
walk the old stack which is both deallocated and not guaranteed to be in
a consistent state. This wasn't reproducible for me as the next function
exit will reset the shadow stack. However, add a fairly complex
functional test anyway that might catch something similar in the future.