compiler: Fix self recursive non-lambda definitions
The optimisation for self recursive lambdas was too eager. If it
detected any self-recursive definition it would disable the logic for
creating a recursive mutable but it wouldn't provide a self value to
replace it.

Fortunately most of the uses of non-lambda self-recursive definitions
aren't legal SCheme. However, if a lambda is embedded in another
expression as is the case for (delay) this bug can be triggered with
legal code.

Also re-enable the last (scheme lazy) test now that this is fixed.