compiler: Clean up ps.Step naming
ps.Step used "Store" to refer to any step that stored a value in a
TempValue that did not fall in to one of the more specific types (e.g.
Unbox*). Additionally, the record-related steps had their verb at the
end instead of the step name which was inconsistent with the other
steps.

- Rename all steps creating a constant value to Create*

- Rename all steps loading a value from memory to Load*

- Rename the record-related steps to be consistent with the other steps
  both in the construction of their names and the verbs used