compiler: Mark GC and error signals as cold
This tells LLVM that _lliby_alloc_cells and _lliby_signal_error are not
called frequently and any code paths leading to them should also be
considered cold. This allows it to generate more efficient code at the
expense of losing support for LLVM 3.3. It also lets us get rid of the
GenLlvmExpect support as its no longer needed

Also mark _lliby_signal_error as noreturn as it unconditionally throws
an exception