Proper named metadata support
Previously we attempted to treat named data the same as "numbered"
metadata. In fact, the two are different and require different syntax.
Rename what we previously called named metadata to "numbered metadata"
and add specific functions for dealing with named metadata

As a test have the compiler emit !llvm.ident metadata with a fixed
identifier string for llambda. More named metadata will be needed in the
future for debugging support.