compiler: Properly located nested data
The Scheme parser was using a single private var to track the start
location of the current datum. However, any nested data would overwrite
the location by the time the data was finished parsing. This moved our
source locations "later" in the input than they should be.

Fix by passing the source location on the value stack. This is cleaner
and more functional anyway.