compiler: Get rid of Scheme identifier regex
The regex was a legacy of the old parser which required it for actual
parsing. It was still used to check for valid identifiers when printing
ASTs so it was retained after the parser rewrite. However, it's much
cleaner and likely faster to just run our parser on the identifier and
see if it parses as a symbol.