compiler: Fix deadlock on startup
By default Scala objects are initialised lazily. This presented a
problem for ct.CellType because the type objects referenced both their
subtypes and supertype. If two threads initialised different CellType
instances it's possible for the threads to deadlock waiting for each
other's CellType instances to initalise

Fix this by getting rid of ct.CellType.supertype completely - we can
just as easily work with just the subtypes. This means the lazy
initialisation always happens in a single direction from superclass =>
subclass regardless of which type is entered first

This fixes a fairly frequent hang in the compiler's test suite