Implement  let:, let*:, letrec:, letrec*:
There are essentially copies of their respective versions in the stdlib.
Their forms are based off of Typed Racket with the exception that types
are always mandatory. This may change at a later time.

This also fixes a planner bug uncovered by letrec:that prevented typed
recursive mutables from compiling.