compiler: Don't treat vt.Predicate as an integer
vt.Predicate is intended to represent boolean values returned from LLVM
tests. However, it was treated as an integer which means it would
convert freely to and from Scheme exact integers.

This had no obvious affect as vt.Predicate is not accessible as an NFI
type and its usages internally are fairly restricted. However, it could
be a source of errors in the future.

Introduce a new vt.BoolLikeType which covers both vt.Predicate and
vt.CBool. Teach TempValueToIntermediate about iv.NativePredicateValue so
it can properly create predicates from temp values