compiler: Introduce RetypeLambdaArgs
RetypeLambdaArgs is a replacement for RefineArgumentTypes. While Refine
inspected the plan steps to deduce types, Retype inspects the reduced
lambda expression using the live intermediate values to determine the
signature of called procedures. This means that Retype works well with
the new Scheme type system as it still has access to the full Scheme
types of the casting expressions while the steps only know about cell
types.

- Introduce RetypeLambdaArgs which inspects the reduced expression of a
  lambda to determine argument types. Rename the RefineArgumentTypes
  test to RetypeLambdaArgs and add a few more test cases to cover

- Have PlanLambda call RetypeLambdaArgs early to rewrite its argument
  types

- Rename the world pointer logic from RefineArgumentTypes to new
  WorldPtrUsedBySteps and use that to drop the world pointer after
  function planning as before

- Remove CastCellToSubtypeChecked and the code in CellValue that went
  out its way to generate it instead of using GenTypeCheck.

- Remove CellType.genTypeCheck. This is no longer usedd