compiler: Support for inlining record procedures
This allows inlining record procedures (constructors, accessors and
predicates) directly in the the plan of the invoking procedure. This
means records are now as efficient as pairs with the added option of
strong typing.

This more than reverses the performance loss from switching promises to
records. It will also allow the upcoming mutable/immutable pair split to
happen without any appreciable performance loss.