compiler: Add range metadata for record class IDs
If we're dealing with a union of possible record types we can pass
through those constraints to LLVM. This is potentially useful as the
record class ID field is 32bits wide but in most programs that range
will be very sparse.

Unfortunately we can't produce ranges when testing for types that aren't
unions of specific record types. That is because record class IDs are
lazily assigned by TypeGenerator during codegen so we don't know the
possible record class IDs up front.