compiler: Introduce constant boolean types
These are specific types for #t and #f. The primary importance of this is
our type system can now implicitly track the truthiness of values. This
means, for example, that (if) can be treated like type predicate for #f
and propagate that truthiness information in to its branches without any
special handling by the planner.

(not) can already be replaced with a type predicate for #f here. This
generates slightly better IR and doesn't require any explicit runtime
or reportproc planner support.