compiler: Remove the reducer
The reducer was complex but worked well enough with the limited
information it had before the introduction of explicit typing. However,
to fully resolve types it would require nearly as much information as
the planner has. This doesn't scale very well.

In the future the type system and planner intermediates should be able
to express much of the information the reducer was tracking (e.g. if
something is a proper list of a given length). This means the planner
can accomplish the same things the reducer was doing at plan time. For
now remove the reducer completely to reduce the complexity of the
compiler while working towards that goal.