compiler: Add static branch logic to planner
This builds on top of the new constant boolean types for #t/#f to do
static type checks on (if) conditions and only planning the taken branch
when possible. This is the first (and likely easiest) step in replacing
the functionality of the reducer

The most interesting part of this is the assertStaticPlan function. This
attempts to assert that a plan collapses down to a single compile-time
determined value. This should be useful for replacing the previous
reducer tests which attempted to do something similar at the expression
level.