compiler: Simple inlining for bound procedures
This introduces a new codegen.CostForPlanSteps function that attempts to
calculate an approximate cost for this plan. The values are fairly
arbitrary and untuned at the moment but it should be significantly
better than the old reducer's heuristics.

Whenever a lambda is planned we attach the original lambda expr and some
other metadata to a new KnownSchemeProc intermediate value. This gives
the planner enough information to re-plan the procedure inline the next
time it's encountered.

Inside PlanApplication the invoke and inline plans are separately
planned in two forks of the PlanWriter. The cost for both plans is
calculated and the cheaper plan is used.