compiler: Static equivalence planning
This re-introduces the unit tests from the reducer and implements static
equivalence planning for constant values. This is more powerful than the
reducer because we can infer non-constant values are not (eqv?) when
their types are disjoint.

This also fixes a typo that was causing (equal?) to not be planned at
all.