compiler: Extend DisposeValues to dispose steps
This replaces the old MergeableStep trait with a new "DisposableStep"
trait. "Disposable" in this case refers to both dropping the step
completely or merging it with other instances of itself. It also adds a
new NullipotentStep trait for steps that are completely nullipotent.

If DisposeValues encounters a nullipotent step with an unused result it
will drop it from the plan completely. I'm guessing LLVM is smart enough
to remove the code itself in the vast majority of cases but this cleans
up our emitted IR and makes the static plan tests easier to write.