compiler: Add codegen constant cache
This creates a module-wide cache of constant cell values created by
codegen. At -O2 this will rarely make a difference due to LLVM's
constant merging. However, this fixes a real semantic problem on -O0
where the lazy instantiation of constant values can cause duplicate
values to be created for the same storage location. This means (eqv?) on
the exact same storage location would return #f

This also has the benefit of making static (memv) and (member) easier to
implement as it won't have to worry about returning a duplicate list
tail when a value is found.