compiler: Static (memv) and (member) planning
This builds on top of the constant cache work to support (memv) and
(member) on constant lists. This is more powerful than the reducer
version in a couple of ways:

1) Even if the needle is non-constant its type constraints can be used
   to determine if it's non-matching. For example, a non-constant value
   of type <integer> couldn't appear in a list of symbols

2) The reducer was strict about the list having to appear inline in the
   call to (memv)/(member) to avoid duplicating list tails. The constant
   cache lifts this restriction and allows us to operate on any known
   constant list.

This also adds a unit test for (case) which is built on top of (memv).
It's equally a test of the inliner but that seems to have been
sufficiently robust to handle (case) without modifications.