compiler: Support for inlining with rest args
This uses the new ValuesToProperList functionality to inject a rest arg
value in to the inline function's body. This is most effective with
immutable pairs as the rest arg's structure and members will be known at
compile time. This allows static planning to occur in many cases.

However, with mutable pairs static planning can occur as long as the
rest arg is treated like an opaque value (passed to another procedure,
returned directly, etc).