compiler: Support inlining lambdas with captures
If a lambda is capturing a variable that the inline site state has
access to we can just use that value directly.

This catches a common pattern where helper produres are defined inside
the body of a larger procedure and then invoked from within that larger
procedure. Previously we would refuse to inline the helper procedures
even if they were capturing values clearly available to the outer
procedure.

This reduces the amount of generated IR for the R7RS life example
program by over 40% and increases its performance by approximately the
same amount.