compiler: Speed up boolean value checks
Because our boolean values are preconstructed we don't need to test if
they're a boolean cell first. We can directly compare the passed cell to
the preconstructed boolean value we're interested in

To accomplish this the derived type infrastructure in PlanTypeCheck has
changed from always testing parent types to allowing individual types to
opt-in. This will become more important as our type hierarchy gets
deeper and many intermediate types won't need to be checked.