compiler: Introduce vt.SatisfiesType
This replaces the vt.SchemeType.satisfiesType method. The logic was
fairly convoluted and spread throughout the various data types in
ValueType.scala.

This moves all the logic in to a separate SatisfiesType function that
takes the two types as arguments. This is easier to follow and will
allow more complicated (and possibly stateful) type checking in the
future.