compiler: Remove empty arg list hack in planner
Previously there was a ps.BuildProperList plan step that would take a
world pointer and build a proper list in codegen. There was a hack in
PlanInvokeApply to avoid creating BuildProperList for empty lists to
prevent a false worldPtr usage

Now we explicitly plan out proper list construction which means we can
directly call ValuesToProperList and it will do the right thing.