compiler: Make report procs return intermediates
Previously ReportProcPlanner was defined to take a PlanResult and every
report proc planner was returning a PlanResult with the initial state
and the value they way to return. It shouldn't be possible for a report
proc to modify the plan state so just returning a value makes more sense
and results in smaller code.