compiler: Don't reuse symbols for native functions
In our analyser track the symbols of all the defined native functions.
Then when we constructor our initial PlanWriter mark all the native
function symbols as allocated. This prevents a symbol collision if a
Scheme procedure with the same name is created. This is needed for the
next commit to support (cos) etc becoming Scheme functions.