Remove implicit integer -> flonum conversion
Previously exact integer would be implicitly converted to inexact
rationals on demand. This has a few problems:

- Exact integers don't satisfy the inexact rational type - they're
  completely disjoint. The fact that they will convert anyway isn't
  discoverable by the type system.

- Unlike (inexact) no provisions were made for checking overflow

- The code to handle conversions was complex and involved phi'ing cell
  values in the worst case

Switch to making exact integer and inexact rational not freely
convertable. If an attempted conversion between number types in
encountered we suggest the use of (exact) or (inexact).

To soften the performance blow simple static planning for
(exact) and (inexact) has been added.