compiler: Proper lists don't satisfy pairs
Proper lists are effectively a union of a special recursive pair type
and the empty list. For that reason it's not possible for them to
definitely satisfy any pair type - they could be the empty list.

Fix by only returning None or Some(false) when checking if a proper
list satisfies a pair. This fixes a crash in an upcoming commit where
the type system was optimising away a check for a proper list being a
pair.