compiler: Support for typed native rest args
This allows native functions to take a uniform list of rest arguments.
This has the primary benefit of supplying more static information to the
compiler. For example, illegal applications such as (+ 1 2 'symbol) are
now caught at compile time.

Currently the runtime will still recheck and recount the argument list.
This is now unnecessary and will be removed shortly.