Rename (ann) => (cast)
(ann) was originally taken from Typed Racket. There, it statically
ensures the given expression has the specified type. If the type
requirement can't be statically satisfied the compile fails. It has
another syntax, (cast), that will make a runtime contract instead of
failing the compile.

Llambda's implementation of (ann) behaved identically to (cast). The
functionality of (ann) is desirable as it allows a way to opt-in to
compile time type guarantees. It also will make writing unit tests for
type deductions much easier.

To prepare for proper (ann) support rename the existing (ann) to (cast)
to free up the name.