compiler: Fix ProperListType static type logic
ProperListType in many ways behaves as a union of the empty list and a
recursive pair type. However, treating it exactly as such causes
infinite recursion in SatisfiesType. For now patch around it by adding
many more proper unit tests and hacking around misbehaviour.

In the future proper recursive types should be supported any
ProperListType should be redefined in terms of that.