compiler: Mark rest arguments as proper lists
Our ABI guarantees that all rest arguments are proper lists. If we have
immutable pairs this means that they must remain proper lists
indefinitely. This means we can internally annotate them as proper list
which triggers additional optimisations with type checking, (length) and
(apply).

This generates signficantly simplified code for (case-lambda) as it can
now calculate list lengths and perform (apply) inline.

This doesn't allow arbitrary list types to be defined in Scheme but is
an intitial step towards that.