compiler: Fix omitting world ptr for Scheme procs
The logic for stripping the world pointer for Scheme procedures was
completely broken. It wouldn't leave the world pointer as a named
argument and in the procedure's signature which means it would cause a
GC safe point every time a Scheme procedure was applied.

Fix PlanLambda and add a simple unit test to ensure trivial Scheme
procedures don't have world pointers.