compiler: Generate cleaner type checks for unions
Change branchOnType and if the true/value branch return static
true/false respectively it will skip the phi completely. This improves
the code we generate for testing union types by removing a useless phi
for the last member in the union. This doesn't affect code generation
but makes the type predicate IR easier to read.