compiler: Support typed rest arguments
This introduces typed rest arguments for Scheme procedures using
identical syntax to Typed Racket. Only uniform lists are supported at
the moment.

This also fixes an issue where the storage location for the rest
argument was defined to have AnySchemeType instead of a more specific
type. This could potentially confuse parts of the compiler in to
thinking the rest arg has a less specific type than it does.

Closes #28